Moved language setting to AppearanceTab
