+++ src/main/java/org/jabref/gui/preferences/appearance/AppearanceTab.fxml
@@ -3,13 +3,17 @@
 <?import javafx.geometry.Insets?>
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.control.CheckBox?>
+<?import javafx.scene.control.ComboBox?>
 <?import javafx.scene.control.Label?>
 <?import javafx.scene.control.RadioButton?>
 <?import javafx.scene.control.Spinner?>
 <?import javafx.scene.control.TextField?>
 <?import javafx.scene.control.ToggleGroup?>
 <?import javafx.scene.control.Tooltip?>
+<?import javafx.scene.layout.ColumnConstraints?>
+<?import javafx.scene.layout.GridPane?>
 <?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.RowConstraints?>
 <?import javafx.scene.layout.VBox?>
 <?import org.jabref.gui.icon.JabRefIconView?>
 <fx:root spacing="10.0" type="VBox"
@@ -19,16 +23,38 @@
         <ToggleGroup fx:id="theme"/>
     </fx:define>
     <Label styleClass="titleHeader" text="%Appearance"/>
+    <GridPane hgap="10.0" vgap="4.0">
+        <columnConstraints>
+            <ColumnConstraints hgrow="SOMETIMES" percentWidth="20.0"/>
+            <ColumnConstraints hgrow="SOMETIMES"/>
+        </columnConstraints>
+        <rowConstraints>
+            <RowConstraints minHeight="10.0" vgrow="SOMETIMES"/>
+            <RowConstraints minHeight="10.0" vgrow="SOMETIMES"/>
+            <RowConstraints minHeight="10.0" vgrow="SOMETIMES"/>
+        </rowConstraints>
 
-    <Label styleClass="sectionHeader" text="%Font"/>
-    <CheckBox fx:id="fontOverride" text="%Override default font settings"/>
-    <HBox alignment="CENTER_LEFT" spacing="4.0">
-        <Label text="%Size" disable="${!fontOverride.selected}"/>
-        <Spinner fx:id="fontSize" styleClass="fontsizeSpinner" editable="true"/>
-        <padding>
-            <Insets left="20.0"/>
-        </padding>
-    </HBox>
+        <Label text="%Language"
+               GridPane.columnIndex="0" GridPane.columnSpan="2"/>
+        <ComboBox fx:id="language" prefWidth="200.0"
+                  GridPane.rowIndex="0" GridPane.columnIndex="1"/>
+
+        <CheckBox fx:id="fontOverride" text="%Override default font settings"
+                  GridPane.rowIndex="1" GridPane.columnIndex="0" GridPane.columnSpan="2"/>
+
+        <Label text="%Size" disable="${!fontOverride.selected}"
+               GridPane.rowIndex="2" GridPane.columnIndex="0">
+            <padding>
+                <Insets left="20.0"/>
+            </padding>
+        </Label>
+        <Spinner fx:id="fontSize" styleClass="fontsizeSpinner" editable="true"
+                 GridPane.rowIndex="2" GridPane.columnIndex="1">
+            <padding>
+                <Insets top="-4.0"/>
+            </padding>
+        </Spinner>
+    </GridPane>
 
     <Label styleClass="sectionHeader" text="%Visual theme"/>
     <RadioButton fx:id="themeLight" text="%Light theme" toggleGroup="$theme"/>
