+++ src/main/java/org/jabref/gui/preferences/appearance/AppearanceTabViewModel.java
@@ -1,19 +1,31 @@
 package org.jabref.gui.preferences.appearance;
 
+import java.util.ArrayList;
+import java.util.Comparator;
+import java.util.List;
+
 import javafx.beans.property.BooleanProperty;
+import javafx.beans.property.ListProperty;
+import javafx.beans.property.ObjectProperty;
 import javafx.beans.property.SimpleBooleanProperty;
+import javafx.beans.property.SimpleListProperty;
+import javafx.beans.property.SimpleObjectProperty;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
+import javafx.collections.FXCollections;
+import javafx.collections.transformation.SortedList;
 import javafx.scene.control.SpinnerValueFactory;
 
 import org.jabref.gui.DialogService;
 import org.jabref.gui.preferences.PreferenceTabViewModel;
 import org.jabref.gui.theme.Theme;
 import org.jabref.gui.util.FileDialogConfiguration;
+import org.jabref.logic.l10n.Language;
 import org.jabref.logic.l10n.Localization;
 import org.jabref.logic.util.StandardFileType;
 import org.jabref.model.strings.StringUtil;
 import org.jabref.preferences.AppearancePreferences;
+import org.jabref.preferences.GeneralPreferences;
 import org.jabref.preferences.PreferencesService;
 
 import de.saxsys.mvvmfx.utils.validation.CompositeValidator;
@@ -23,10 +35,11 @@ import de.saxsys.mvvmfx.utils.validation.ValidationStatus;
 import de.saxsys.mvvmfx.utils.validation.Validator;
 
 public class AppearanceTabViewModel implements PreferenceTabViewModel {
-
-    public static SpinnerValueFactory<Integer> fontSizeValueFactory =
+    protected static SpinnerValueFactory<Integer> fontSizeValueFactory =
             new SpinnerValueFactory.IntegerSpinnerValueFactory(9, Integer.MAX_VALUE);
 
+    private final ListProperty<Language> languagesListProperty = new SimpleListProperty<>();
+    private final ObjectProperty<Language> selectedLanguageProperty = new SimpleObjectProperty<>();
     private final BooleanProperty fontOverrideProperty = new SimpleBooleanProperty();
     private final StringProperty fontSizeProperty = new SimpleStringProperty();
     private final BooleanProperty themeLightProperty = new SimpleBooleanProperty();
@@ -37,14 +50,18 @@ public class AppearanceTabViewModel implements PreferenceTabViewModel {
     private final DialogService dialogService;
     private final PreferencesService preferences;
     private final AppearancePreferences appearancePreferences;
+    private final GeneralPreferences generalPreferences;
 
     private final Validator fontSizeValidator;
     private final Validator customPathToThemeValidator;
+
+    private final List<String> restartWarning = new ArrayList<>();
 
     public AppearanceTabViewModel(DialogService dialogService, PreferencesService preferences) {
         this.dialogService = dialogService;
         this.preferences = preferences;
         this.appearancePreferences = preferences.getAppearancePreferences();
+        this.generalPreferences = preferences.getGeneralPreferences();
 
         fontSizeValidator = new FunctionBasedValidator<>(
                 fontSizeProperty,
@@ -71,6 +88,8 @@ public class AppearanceTabViewModel implements PreferenceTabViewModel {
 
     @Override
     public void setValues() {
+        languagesListProperty.setValue(new SortedList<>(FXCollections.observableArrayList(Language.values()), Comparator.comparing(Language::getDisplayName)));
+        selectedLanguageProperty.setValue(generalPreferences.getLanguage());
         fontOverrideProperty.setValue(appearancePreferences.shouldOverrideDefaultFontSize());
         fontSizeProperty.setValue(String.valueOf(appearancePreferences.getMainFontSize()));
 
@@ -95,6 +114,13 @@ public class AppearanceTabViewModel implements PreferenceTabViewModel {
 
     @Override
     public void storeSettings() {
+        Language newLanguage = selectedLanguageProperty.getValue();
+        if (newLanguage != generalPreferences.getLanguage()) {
+            generalPreferences.setLanguage(newLanguage);
+            Localization.setLanguage(newLanguage);
+            restartWarning.add(Localization.lang("Changed language") + ": " + newLanguage.getDisplayName());
+        }
+
         appearancePreferences.setShouldOverrideDefaultFontSize(fontOverrideProperty.getValue());
         appearancePreferences.setMainFontSize(Integer.parseInt(fontSizeProperty.getValue()));
 
@@ -134,6 +160,19 @@ public class AppearanceTabViewModel implements PreferenceTabViewModel {
             return false;
         }
         return true;
+    }
+
+    @Override
+    public List<String> getRestartWarnings() {
+        return restartWarning;
+    }
+
+    public ListProperty<Language> languagesListProperty() {
+        return this.languagesListProperty;
+    }
+
+    public ObjectProperty<Language> selectedLanguageProperty() {
+        return this.selectedLanguageProperty;
     }
 
     public BooleanProperty fontOverrideProperty() {
