+++ src/main/java/org/jabref/gui/preferences/library/LibraryTab.java
@@ -7,7 +7,6 @@ import javafx.scene.control.ComboBox;
 import org.jabref.gui.preferences.AbstractPreferenceTabView;
 import org.jabref.gui.preferences.PreferencesTab;
 import org.jabref.gui.util.ViewModelListCellFactory;
-import org.jabref.logic.l10n.Language;
 import org.jabref.logic.l10n.Localization;
 import org.jabref.model.database.BibDatabaseMode;
 
@@ -15,7 +14,6 @@ import com.airhacks.afterburner.views.ViewLoader;
 
 public class LibraryTab extends AbstractPreferenceTabView<LibraryTabViewModel> implements PreferencesTab {
 
-    @FXML private ComboBox<Language> language;
     @FXML private ComboBox<BibDatabaseMode> biblatexMode;
     @FXML private CheckBox inspectionWarningDuplicate;
     @FXML private CheckBox confirmDelete;
@@ -37,12 +35,6 @@ public class LibraryTab extends AbstractPreferenceTabView<LibraryTabViewModel> i
 
     public void initialize() {
         this.viewModel = new LibraryTabViewModel(dialogService, preferencesService.getGeneralPreferences(), preferencesService.getTelemetryPreferences());
-
-        new ViewModelListCellFactory<Language>()
-                .withText(Language::getDisplayName)
-                .install(language);
-        language.itemsProperty().bind(viewModel.languagesListProperty());
-        language.valueProperty().bindBidirectional(viewModel.selectedLanguageProperty());
 
         new ViewModelListCellFactory<BibDatabaseMode>()
                 .withText(BibDatabaseMode::getFormattedName)
